+++ packages/dumb/src/modules/print/settings-print-view/settings-print-view.util.ts
@@ -169,17 +169,11 @@ const getTableRowsByDataTableType = (type: PdfSettingsDataType, data: TableData,
           label: t('Cgm sensor expiration date'),
           value: cgmData.expirationDate ? formatLocalizedFromUTC(cgmData.expirationDate, timePrefs, longDayFormat) : ''
         }, {
-          label: t('pairing-code'),
-          value: cgmData.pairingCode || ''
-        }, {
           label: t('serial-number'),
           value: cgmData.serialNumber || ''
         }, {
           label: t('software-number'),
           value: cgmData.softwareNumber || ''
-        }, {
-          label: t('Cgm transmitter id'),
-          value: cgmData.firmwareVersion || ''
         }]
       }